Move epic of publication on over to github epic
